# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requirements?: string;
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  skills: string[];
  experienceYears: number;
  education: string;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  skillMatchScore?: number | null;
  /** @nullable */
  experienceMatchScore?: number | null;
  /** @nullable */
  aiRecommendation?: string | null;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  jobId?: number | null;
  createdAt: string;
}

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  skills: string[];
  experienceYears: number;
  education: string;
  resumeText?: string;
  /** @nullable */
  jobId?: number | null;
}

export interface CandidateUpdate {
  name?: string;
  email?: string;
  skills?: string[];
  experienceYears?: number;
  education?: string;
  resumeText?: string;
}

export interface CsvImportInput {
  csvData: string;
  /** @nullable */
  jobId?: number | null;
}

export interface CsvImportResult {
  imported: number;
  errors: string[];
}

export interface RankInput {
  jobId: number;
}

export interface AnalyzeInput {
  jobId: number;
}

export interface AnalyticsSummary {
  totalCandidates: number;
  totalJobs: number;
  analyzedCandidates: number;
  avgScore: number;
  topCandidates: Candidate[];
}

export interface SkillCount {
  skill: string;
  count: number;
}

export interface ScoreBucket {
  range: string;
  count: number;
}

export type ListCandidatesParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type GetAnalyticsSummaryParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type GetSkillDistributionParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type GetScoreDistributionParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type ExportCsvParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type ExportXlsxParams = {
/**
 * @nullable
 */
jobId?: number | null;
};

export type ExportPdfParams = {
/**
 * @nullable
 */
jobId?: number | null;
};