# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TalentMind AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Add a candidate manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates/import-csv:
    post:
      operationId: importCandidatesCsv
      tags: [candidates]
      summary: Import candidates from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvImportResult"
  /analysis/rank:
    post:
      operationId: rankCandidates
      tags: [analysis]
      summary: Rank all candidates for a job using Gemini AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankInput"
      responses:
        "200":
          description: Ranked candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
  /analysis/candidate/{id}:
    post:
      operationId: analyzeCandidate
      tags: [analysis]
      summary: Analyze a single candidate against a job using Gemini AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInput"
      responses:
        "200":
          description: Analyzed candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard analytics summary
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/skill-distribution:
    get:
      operationId: getSkillDistribution
      tags: [analytics]
      summary: Get top skills distribution across candidates
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Skill distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillCount"
  /analytics/score-distribution:
    get:
      operationId: getScoreDistribution
      tags: [analytics]
      summary: Get score distribution buckets
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Score distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreBucket"
...
components:
  schemas:
...
    Candidate:
      type: object
      required: [id, name, email, skills, experienceYears, education, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: number
        education:
          type: string
        resumeText:
          type: ["string", "null"]
        aiScore:
          type: ["number", "null"]
        skillMatchScore:
          type: ["number", "null"]
        experienceMatchScore:
          type: ["number", "null"]
        aiRecommendation:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        rank:
          type: ["integer", "null"]
        jobId:
          type: ["integer", "null"]
        createdAt:
          type: string
    CandidateInput:
      type: object
      required: [name, email, skills, experienceYears, education]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: number
        education:
          type: string
        resumeText:
          type: string
        jobId:
          type: ["integer", "null"]
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: number
        education:
          type: string
        resumeText:
          type: string
    CsvImportInput:
      type: object
      required: [csvData]
      properties:
        csvData:
          type: string
        jobId:
          type: ["integer", "null"]
    CsvImportResult:
      type: object
      required: [imported, errors]
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string
    RankInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
    AnalyzeInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
    AnalyticsSummary:
      type: object
      required: [totalCandidates, totalJobs, analyzedCandidates, avgScore, topCandidates]
      properties:
        totalCandidates:
          type: integer
        totalJobs:
          type: integer
        analyzedCandidates:
          type: integer
        avgScore:
          type: number
        topCandidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
    SkillCount:
      type: object
      required: [skill, count]
      properties:
        skill:
          type: string
        count:
          type: integer
    ScoreBucket:
      type: object
      required: [range, count]
      properties:
        range:
          type: string
        count:
          type: integer