# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from './candidate';

export interface AnalyticsSummary {
  totalCandidates: number;
  totalJobs: number;
  analyzedCandidates: number;
  avgScore: number;
  topCandidates: Candidate[];
}