# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  id: number;
  name: string;
  email: string;
  skills: string[];
  experienceYears: number;
  education: string;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  skillMatchScore?: number | null;
  /** @nullable */
  experienceMatchScore?: number | null;
  /** @nullable */
  aiRecommendation?: string | null;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  jobId?: number | null;
  createdAt: string;
}