# lib/api-zod/src/generated/types/candidateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  skills: string[];
  experienceYears: number;
  education: string;
  resumeText?: string;
  /** @nullable */
  jobId?: number | null;
}