# lib/api-zod/src/generated/types/csvImportInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CsvImportInput {
  csvData: string;
  /** @nullable */
  jobId?: number | null;
}