# lib/api-zod/src/generated/types/csvImportResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TalentMind AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CsvImportResult {
  imported: number;
  errors: string[];
}